Add support for changing currency in Settings
Stripe is now available in Canada so you are now able to choose to process in either American ("usd") and Canadian ("cad") currency.

At the moment, "usd" is hard-coded as the currency variable in several spots which causes errors if your account is set to "cad". These changes allow the currency to be set as one of the variables in PAYMENT_PLANS.