Extend the stripe_id fields to 255 characters
There was an announcement on the Stripe API mailing
list that they would be extending beyond 30 chars
soon.

Stripe hadn't previously picked a hard limit that
was published externally but after looking at IDs
I decided to make it a bit bigger than what I was
seeing and picked an arbitrary "50" for the length.

Stripe has committed to keeping it less than 50 for
at least the next 6 months and beyond that they are
committing to keeping them less than 255 characters.

See http://bit.ly/19cu4w8 for more details.