Blank lines should not contain any whitespace
Eldarion has decided for open source projects to conform to the blank lines
should not contain whitespace style guideline. We want to do less style guide enforcing with our contributors and make it easier for them to contribute.

This does not mean we won't enforce this style guideline. Consistency is still
king and mixing whitespace lines with no whitespace lines is discouraged.