Remove context processor
This is a backwards incompatible change in that if
you have the processor installed in settings.py
you will need to remove it, however, all views now
include this context.

If you have written you own custom views and are
relying on the context processor to put values in
context, you'll need to update your views.