Add tests for actions.accounts
Note: I'm not 100% sure what the valid data responses are for custom accounts. Based on the branching I saw in sync_account_from_stripe_data it appears that there are expected conditions where the data doesn't have some values. However, when looking at the API docs it appears these are not optional data elements. If that is the case we can remove the extra test, data, and remove the branching in the sync function